Create permissions and files modules
